api_layer/requirements.txt
aiohttp==3.13.3
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astroid==4.0.3
asyncpg==0.31.0
azure-core==1.38.0
azure-identity==1.25.1
azure-keyvault-secrets==4.10.0
azure-storage-blob==12.28.0
black==25.12.0
build==1.4.0
certifi==2026.1.4
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
coverage==7.13.1
cryptography==46.0.3
databricks-sdk==0.77.0
dill==0.4.0
distlib==0.4.0
docutils==0.22.4
fastapi==0.128.0
filelock==3.20.2
flake8==7.3.0
flake8-docstrings==1.7.0
Flake8-pyproject==1.2.4
google-auth==2.47.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
id==1.5.0
identify==2.6.15
idna==3.11
iniconfig==2.3.0
isodate==0.7.2
isort==7.0.0
jaraco.classes==3.4.0
jaraco.context==6.1.0
jaraco.functools==4.4.0
jeepney==0.9.0
keyring==25.7.0
librt==0.7.7
loguru==0.7.3
mando==0.7.1
markdown-it-py==4.0.0
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.8.0
msal==1.34.0
msal-extensions==1.3.1
mypy==1.19.1
mypy_extensions==1.1.0
nh3==0.3.2
nodeenv==1.10.0
packaging==25.0
pathspec==1.0.2
platformdirs==4.5.1
pluggy==1.6.0
pre_commit==4.5.1
protobuf==6.33.2
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydocstyle==6.3.0
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pylint==4.0.4
pyproject_hooks==1.2.0
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
python-dotenv==1.2.1
pytokens==0.3.0
PyYAML==6.0.3
radon==6.0.1
readme_renderer==44.0
requests==2.32.5
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.2.0
rsa==4.9.1
SecretStorage==3.5.0
six==1.17.0
snowballstemmer==3.0.1
starlette==0.50.0
tomlkit==0.13.3
twine==6.2.0
types-requests==2.32.4.20260107
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.40.0
virtualenv==20.36.0
